requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
setuptools==69.2.0
sqlparse==0.4.4
